backend/requirements.txt
Flask==1.1.1
flask-restful==0.3.8
flask-cors==3.0.7
flask_sqlalchemy==2.3
sqlalchemy-utils==0.36.1
flask-sqlalchemy-session==1.1
psycopg2-binary
alembic==1.4.0
bcrypt==3.1.7
pyjwt==1.7.1
